app/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
sqlalchemy
cryptography